requirements.txt
fastapi
uvicorn[standard]
python-dotenv
ebooklib
beautifulsoup4
httpx
lxml
python-multipart